// Frontend/generated/models/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Backend | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Department } from './department';

export interface User {
  id?: string;
  /** @nullable */
  entraUserId?: string | null;
  /** @nullable */
  role?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  departmentId?: string | null;
  department?: null | Department;
}